Add CalendarView, AnalysisView, Bottom Buttons
Both CalendarView and AnalysisVIew are embedded in large screen view that tagged on 1. Tabview can make those kind of screens moved on and this can be implemented with finger swipe gesture. Or It can be controlled by three buttons at the bottom of the screen. However, the feature of button isn't available yet and will be available in the next update.